+++ database/migrations/2021_09_10_150223_exchange_collectibles.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class ExchangeCollectibles extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('exchange_collectibles', function (Blueprint $table) {
+            $table->id();
+            $table->unsignedBigInteger('classroom_id');
+            $table->unsignedBigInteger('student_id');
+            $table->unsignedBigInteger('collectionable_id');
+            $table->unsignedBigInteger('wanted_collectionable_id');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('exchange_collectibles');   
+    }
+}