+++ resources/lang/ca/notifications.php
@@ -29,5 +29,6 @@ return [
     'post' => 'Nou post en: ',
     'mark_challenge' => 'Ha completat un repte: ',
     'new_collectionable' => 'Nou col·leccionable',
+    'challenge_password' => 'Ha superat el repte amb contrasenya. ',
 
 ];