+++ resources/lang/ca/success_error.php
@@ -46,4 +46,6 @@ return [
     'max_cards_reached' => 'Màxim de cartes superat!',
     
     'craft_fail' => 'Aquests objectes no es poden combinar',
+    'collectible_error' => 'No es compleixen els requeriments d\'intercanvi: tenir mínim 2 coleccionables. Comprova que el company o companya té els 2 col·leccionables disponibles.',
+
 ];