+++ resources/lang/en/notifications.php
@@ -28,6 +28,7 @@ return [
     'post' => 'New post on: ',
     'mark_challenge' => 'Has completed one challenge: ',
     'new_collectionable' => 'New collectible',
+    'challenge_password' => 'Has passed a password challenge. ',
 
 
 ];