+++ resources/lang/en/success_error.php
@@ -45,6 +45,8 @@ return [
     'max_cards_reached' => 'Max cards reached!',
 
     'craft_fail' => 'These objects can\'t be crafted',
+    'collectible_error' => 'Exchange requirements are not met: have at least 2 collectibles. Check if your mate has the 2 collectibles available.',
+
 
 
 ];