+++ resources/lang/es/success_error.php
@@ -47,5 +47,6 @@ return [
     'max_cards_reached' => '¡Máximo de cartas superado!',
 
     'craft_fail' => 'Estos objetos no se pueden combinar',
+    'collectible_error' => 'No se cumplen los requisito de intercambio: tener mínimo 2 coleccionables. Comprueba que tu compañero o compañera tiene los 2 coleccionables de ese tipo.',
 
 ];